# API/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.AtualizarProntuarioRequest:
    properties:
      conteudo:
        type: string
      titulo:
        type: string
    required:
    - conteudo
    - titulo
    type: object
...
  controllers.CriarProntuarioRequest:
    properties:
      conteudo:
        example: Paciente evolui bem.
        type: string
      paciente_id:
        example: 1
        type: integer
      titulo:
        example: Evolução
        type: string
    required:
    - conteudo
    - paciente_id
    - titulo
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        example: Erro na operação
        type: string
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        example: Operação realizada com sucesso
        type: string
    type: object
...
  controllers.ProcedimentoRequest:
    properties:
      descricao:
        example: Consulta médica geral
        type: string
      duracao:
        example: 30
        type: integer
      nome:
        example: Consulta Médica
        type: string
      valor:
        example: 150
        type: number
    required:
    - nome
    - valor
    type: object
  controllers.ProcedimentoResponse:
    properties:
      ativo:
        example: true
        type: boolean
      clinica_id:
        example: 1
        type: integer
      descricao:
        example: Consulta médica geral
        type: string
      duracao:
        example: 30
        type: integer
      id:
        example: 1
        type: integer
      nome:
        example: Consulta Médica
        type: string
      valor:
        example: 150
        type: number
    type: object
  controllers.ProntuarioRegistroSwagger:
    properties:
      clinica_id:
        example: 1
        type: integer
      conteudo:
        example: '...'
        type: string
      created_at:
        example: "2026-04-05T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      paciente_id:
        example: 1
        type: integer
      profissional_id:
        example: 2
        type: integer
      titulo:
        example: Evolução
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API multiclínica com isolamento por tenant (clinic_id), RBAC por papel
    (ADM_GERAL, DONO, MEDICO, SECRETARIA), agendamento com validação de conflitos,
    prontuário com imutabilidade após 24h, auditoria LGPD e métricas de faturamento
    e no-show.
  title: API White Label — Clínicas
  version: "1.0"
paths:
...
  /clinicas/prontuarios:
    get:
      parameters:
      - description: ID do paciente
        in: query
        name: paciente_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ProntuarioRegistroSwagger'
            type: array
      security:
      - BearerAuth: []
      summary: Listar prontuário por paciente
      tags:
      - Prontuário
    post:
      consumes:
      - application/json
      description: Registra evolução clínica (dados sensíveis — auditoria LGPD).
      parameters:
      - description: Dados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.CriarProntuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.ProntuarioRegistroSwagger'
      security:
      - BearerAuth: []
      summary: Criar registro de prontuário
      tags:
      - Prontuário
  /clinicas/prontuarios/{id}:
    put:
      description: Permitido apenas até 24h após a criação.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Dados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AtualizarProntuarioRequest'
      responses: {}
      security:
      - BearerAuth: []
      summary: Atualizar registro de prontuário
      tags:
      - Prontuário
...
  /procedimentos:
    get:
      consumes:
      - application/json
      description: Lista todos os procedimentos da clínica
      parameters:
      - description: Filtrar apenas procedimentos ativos
        in: query
        name: ativas
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ProcedimentoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar procedimentos por clínica
      tags:
      - Procedimentos
    post:
      consumes:
      - application/json
      description: Cadastra um novo procedimento na clínica
      parameters:
      - description: Dados do procedimento
        in: body
        name: procedimento
        required: true
        schema:
          $ref: '#/definitions/controllers.ProcedimentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cadastrar procedimento
      tags:
      - Procedimentos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"